requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
httpx>=0.27.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
groq>=0.11.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
